refactor: queryClient 분리
