+++ src/App.tsx
@@ -1,10 +1,9 @@
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
+import { QueryClientProvider } from "@tanstack/react-query"
 import { BrowserRouter as Router } from "react-router-dom"
 import PostsManagerPage from "./pages/PostsManagerPage.tsx"
+import { queryClient } from "./shared/api/queryClient.ts"
 import Footer from "./widgets/ui/Footer.tsx"
 import Header from "./widgets/ui/Header.tsx"
-
-const queryClient = new QueryClient()
 
 const App = () => {
   return (
