+++ src/entities/user/api/userApi.ts
@@ -1,4 +1,4 @@
-import { safeFetch } from "../../../shared/lib"
+import { safeFetch } from "../../../shared/api"
 import { UserDTO } from "../model/types"
 
 export const userApi = {
