+++ src/shared/api/queryClient.ts
@@ -1,0 +1,3 @@
+import { QueryClient } from "@tanstack/react-query"
+
+export const queryClient = new QueryClient()
