+++ src/shared/api/safeFetch.ts
